+++ examples/clob/programs/Cargo.toml
@@ -8,11 +8,3 @@ path = "src/lib.rs"
 
 [dependencies]
 clob-sp1 = { path = "sp1" }
-
-[dev-dependencies]
-clob-core = { workspace = true }
-borsh = { workspace = true }
-alloy = { workspace = true, features = ["full", "contract", "transport-http", "transport-ws", "network", "providers", "pubsub", "signer-local", "signer-keystore", "rpc-types-debug", "rpc-types-trace", "json-rpc", "rpc-client", "rlp", "sol-types"] }
-ivm-zkvm = { workspace = true }
-ivm-abi = { workspace = true }
-clob-test-utils = { workspace = true }
