+++ examples/matching-game/programs/Cargo.toml
@@ -8,11 +8,3 @@ path = "src/lib.rs"
 
 [dependencies]
 matching-game-sp1 = { path = "sp1" }
-ivm-zkvm = { workspace = true }
-
-[dev-dependencies]
-matching-game-core = { workspace = true }
-bincode = { workspace = true }
-alloy = { workspace = true, features = ["full", "contract", "transport-http", "transport-ws", "network", "providers", "pubsub", "signer-local", "signer-keystore", "rpc-types-debug", "rpc-types-trace", "json-rpc", "rpc-client", "rlp", "sol-types"] }
-ivm-abi = { workspace = true }
-kairos-trie = { workspace = true }
